Development

+++ src/controllers/movies.controller.js
@@ -1,15 +1,55 @@
 import { Pelicula } from '../models/Pelicula.js';
 import { Comentario } from '../models/Comentarios.js'
+import { sequelize } from '../database/connection.js';
 
-export const getMovies = async (req, res) => {
+export const getTop5RecentMovies = async (req, res) => {
   try {
-    const movies = await Pelicula.findAll();
-    res.json(movies);
+    const [result] = await sequelize.query('CALL ObtenerIdTop5PeliculasMasRecientes()', {
+      type: sequelize.QueryTypes.SELECT,
+    });
+    
+    // Get ids from the result object.
+    const movieIDs = Object.values(result).map(item => item.peliculaID);
+  
+    if (movieIDs.length === 0) {
+      res.status(404).json({message:"No se han encontrado peliculas..."});
+      return;
+    }
+
+    const movieInfoPromises = [];
+
+    for (let i = 0; i < movieIDs.length; i++) {
+      const movieInfoPromise = getMovieInfo(movieIDs[i]);
+      movieInfoPromises.push(movieInfoPromise);
+    }
+
+    const moviesInfo = await Promise.all(movieInfoPromises);
+
+    res.status(200).json(moviesInfo);
+
   } catch (error) {
-    console.log(`An error has ocurred while getting movies: ${error}`);
-    res.status(500).json({message: error.message});
+    console.log(`An error has occurred while getting movies: ${error}`);
+    res.status(500).json({ message: error.message });
   }
 };
+
+export const getMovieInfoById = async (req, res) => {
+  try {
+    const { id } = req.params;
+
+    const fullMovieInfo = await getMovieInfo(id);
+
+    if (fullMovieInfo) {
+      res.status(200).json(fullMovieInfo);
+    } else {
+      res.status(404).json({ message: "No se ha encontrado la película." });
+    }
+  } catch (error) {
+    console.error(`Ha ocurrido un error al obtener información de la película: ${error}`);
+    res.status(500).json({ message: "Error interno del servidor." });
+  }
+};
+
 
 export const createMovie = async (req, res) => {
   try {
@@ -33,13 +73,18 @@ export const createMovie = async (req, res) => {
 export const updateMovie = async (req, res) => {
   try {
     const { id } = req.params;
-    const { nombre, resena, califiacionID, poster, fecha } = req.body;
+    const { nombre, resena, calificacionID, poster, fecha } = req.body;
 
     const movie = await Pelicula.findByPk(id);
 
+    if (!movie) {
+        res.status(404).json({ message: 'La película a modificar no existe.' });
+        return;
+    }
+
     movie.nombre = nombre;
     movie.resena = resena;
-    movie.califiacionID = califiacionID;
+    movie.calificacionID = calificacionID;
     movie.poster = poster;
     movie.fecha = fecha;
 
@@ -57,6 +102,12 @@ export const deleteMovie = async (req, res) => {
   try {
     // Get id from the url parameters.
     const { id } = req.params;
+
+    const movie = await Pelicula.findByPk(id);
+    if (!movie) {
+        res.status(404).json({ message: 'La película a eliminar no existe.' });
+        return;
+    }
     
     // Delete user.
     await Pelicula.destroy({
@@ -98,6 +149,10 @@ export const updateComment = async (req, res) => {
     const { usuarioID, peliculaID, contenido, comentarioPadreID, fecha } = req.body;
 
     const comment = await Comentario.findByPk(id);
+    if (!comment) {
+      res.status(404).json({ message: 'El comentario a modificar no existe.' });
+      return;
+    }
 
     comment.usuarioID = usuarioID;
     comment.peliculaID = peliculaID;
@@ -118,7 +173,12 @@ export const updateComment = async (req, res) => {
 export const deleteComment = async (req, res) => {
   try {
     const { id } = req.params;
-    console.log(id);
+
+    const comment = await Comentario.findByPk(id);
+    if (!comment) {
+      res.status(404).json({ message: 'El comentario a eliminar no existe.' });
+      return;
+  }
 
     await Comentario.destroy({
       where: {
@@ -131,4 +191,42 @@ export const deleteComment = async (req, res) => {
     console.log(`An error has ocurred while deleting the comment: ${error}`);
     res.status(500).json({message: error.message});
   }
+}
+
+
+// Functions
+const isEmpty = (obj) => Object.keys(obj).length === 0;
+
+async function getMovieInfo(id) {
+  const movieInfo = await sequelize.query('CALL ObtenerInformacionPelicula(:peliculaID)', {
+    replacements: { peliculaID: id },
+    type: sequelize.QueryTypes.SELECT,
+  });
+
+  // Validar que encuentra la pelicula.
+  if (isEmpty(movieInfo[0])) return null;
+
+  const calificaciones = await sequelize.query('CALL ObtenerCalificacionesPelicula(:peliculaID)', {
+    replacements: { peliculaID: id },
+    type: sequelize.QueryTypes.SELECT,
+  });
+
+  const involucrados = await sequelize.query('CALL ObtenerInvolucradosPelicula(:peliculaID)', {
+    replacements: { peliculaID: id },
+    type: sequelize.QueryTypes.SELECT,
+  });
+
+  const comentarios = await sequelize.query('CALL ObtenerComentariosPelicula(:peliculaID)', {
+    replacements: { peliculaID: id },
+    type: sequelize.QueryTypes.SELECT,
+  });
+
+  const fullMovieInfo = {
+    movieInfo: movieInfo[0]['0'],
+    calificaciones: calificaciones[0]['0'], 
+    involucrados: involucrados[0] || null,
+    comentarios: comentarios[0] || null,
+  }
+
+  return fullMovieInfo;
 }
